rework python packaging. again
